Refactor AILogic into smaller, more manageable classes
This commit refactors the AILogic class into a set of smaller, more manageable classes. The new architecture uses the State and Strategy patterns to make the AI's logic more extensible and easier to maintain.

The following changes were made:

- The AILogic class was moved to `src/ai_logic/main.py`.
- The `TargetFinder` class was created to handle finding targets.
- The `Explorer` class was created to handle exploring the map.
- The `AIState` class and its subclasses were created to manage the AI's state.
- The `AILogic` class was refactored to use the new classes and manage state.
- The tests were updated to reflect the new class structure.
