+++ src/ai_logic/__init__.py
@@ -1,0 +1,3 @@
+from .main import AILogic
+
+__all__ = ["AILogic"]
